requirements/test.txt
-r base.txt
coverage==7.13.5
flake8==7.3.0
flake8-docstrings==1.7.0
Pillow==12.2.0
pytest==9.0.3
pytest-cov==7.1.0
tox==4.53.0
